core/requirements.txt
asgiref==3.5.0 
django==4.0.2 
sqlparse==0.4.2 
tzdata==2021.5
djangorestframework==3.13.1
pytz==2021.3
cffi==1.15.0 
cryptography==36.0.1
django-rest-knox==4.2.0 
pycparser==2.21